+++ Changes.md
@@ -5,6 +5,7 @@
 HEAD
 ----------
 
+- Backport `Sidekiq::Web.configure` for compatibility with 8.0 [#6532]
 - Fix dead filtering [#6508]
 
 7.3.6